refactor(arch): move InMemoryProviderRegistry to infrastructure layer
Move concrete implementation InMemoryProviderRegistry from domain to
infrastructure layer following Hexagonal Architecture principles.

Changes:
- Create src/bioetl/infrastructure/provider_registry.py with InMemoryProviderRegistry
- Add get_provider_registry() factory function in domain layer
- Maintain backward compatibility via __getattr__ lazy import
- Update imports in infrastructure/config/provider_registry.py
- Update test imports to use new location

The domain layer now only contains abstractions (ProviderRegistryABC)
and errors, while concrete implementations reside in infrastructure.
